Centralized EdgeType enum and related classification functions into a single file. Added extension to return the index of an element in an indexable collection Added class for SliceChords
Many helper functions and improvements to MorphRenderMesh
Added function to indicate if a vertex has a complete set of faces
Added the idea of artificial verticies and edges that are added to close holes that do not connect to other regions
MorphRenderVertex can have a null PointIndex if it was an artificially added point
Fixes for Region detection
Fixes for Region conversion to Polygon and ordering of region verticies sequentialy around the perimeter
EdgeType detection can take orientation of edges into account

Performance:
Added a function parameter to FindFacesInPath that provides hints about whether a face can be part of the path.  Currently one use is
used to ensure all faces are connected to a single vertex.