This checkin is the first mesh generator that successfully generates the hardest test case.
Allow serialization and loading of Polygons for debugging purposes
Added ability for MonogameTestbed Tests to UnloadContent on completion
Added ability to enumerate a MorphRenderFace's adjacent faces restricted to faces that pass a criteria function.
FindCloseableFaces now uses the face with the shortest Euclidean Perimeter when multiple possible faces are found.
Regions use MedialAxis to close untiled regions
Updates to Monogame shaders to be consistent with XNA shader updates.
BajajText now uses lists of views to represent the mesh at various stages of generation to facilitate debugging.